+++ apps/chronicles/navigation/index.tsx
@@ -7,41 +7,28 @@ import React from "react";
 import { ColorSchemeName } from "react-native";
 
 import NotFoundScreen from "../screens/NotFoundScreen";
-import DiscoverScreen from "../screens/DiscoverScreen";
+import DiscoverScreen from "../screens/components/DiscoverScreen";
 import IconsScreen from "../screens/IconsScreen";
-import { DiscoverStackParamList, RootTabParamList } from "../types";
 import LinkingConfiguration from "./LinkingConfiguration";
 import { DrawScreen } from "../screens/DrawScreen";
 import { SignatureScreen } from "../screens/SignatureTest";
-import { PaperScreen } from "../screens/components/PaperScreen";
-import { PopoverScreen } from "../screens/components/PopoverScreen";
-import { ButtonScreen } from "../screens/components/ButtonScreen";
 import { Color, Icon, IconName, useToken } from "@equinor/mad-components";
-import { InputScreen } from "../screens/components/InputScreen";
-import { TextFieldScreen } from "../screens/components/TextFieldScreen";
-import { SearchScreen } from "../screens/components/SearchScreen";
-import { CellScreen } from "../screens/components/CellScreen";
-import { NavigationCellScreen } from "../screens/components/NavigationCellScreen";
-import { AccordionScreen } from "../screens/components/AccordionScreen";
-import { MenuScreen } from "../screens/components/MenuScreen";
-import { ProgressIndicatorScreen } from "../screens/components/ProgressIndicatorScreen";
-import { PortalScreen } from "../screens/components/PortalScreen";
-import { DialogScreen } from "../screens/components/DialogScreen";
-import { EnvironmentScreen } from "../screens/components/EnvironmentScreen";
-import { ButtonCellScreen } from "../screens/components/ButtonCellScreen";
-import { SwitchCellScreen } from "../screens/components/SwitchCellScreen";
-import { SelectionControlsScreen } from "../screens/components/SelectionControlsScreen";
-import { ErrorBoundaryScreen } from "../screens/components/ErrorBoundaryScreen";
 import {
     createBottomTabNavigator,
     createNativeStackNavigator,
     createCoreStackNavigator,
-    NavigationContainer
+    NavigationContainer,
 } from "@equinor/mad-core";
 import { config } from "../mad.config";
-import { AutocompleteScreen } from "../screens/components/AutocompleteScreen";
 import { GoToSettingsButton } from "../components/GoToSettingsButton";
 import { SampleSettingsScreen } from "./SettingsScreen";
+import {
+    ComponentsStackParamList,
+    RootStackParamList,
+    RootTabParamList,
+} from "../types/navigation";
+import { ComponentScreen } from "../screens/components/ComponentScreen";
+import { ComponentName } from "../types/components";
 
 export default function Navigation({ colorScheme }: { colorScheme: ColorSchemeName }) {
     const token = useToken();
@@ -65,7 +52,7 @@ export default function Navigation({ colorScheme }: { colorScheme: ColorSchemeNa
     );
 }
 
-const CoreStack = createCoreStackNavigator(config);
+const CoreStack = createCoreStackNavigator<RootStackParamList>(config);
 function RootNavigator() {
     return (
         <CoreStack.Navigator>
@@ -84,10 +71,10 @@ function RootNavigator() {
     );
 }
 
-const DiscoverStack = createNativeStackNavigator<DiscoverStackParamList>();
+const ComponentsStack = createNativeStackNavigator<ComponentsStackParamList>();
 function DiscoverNavigator() {
     return (
-        <DiscoverStack.Navigator
+        <ComponentsStack.Navigator
             initialRouteName="Discover"
             screenOptions={{
                 headerLargeTitle: true,
@@ -101,35 +88,16 @@ function DiscoverNavigator() {
                 headerRight: () => <GoToSettingsButton marginRight={-12} />,
             }}
         >
-            <DiscoverStack.Screen name="Discover" component={DiscoverScreen} />
-            <DiscoverStack.Screen name="Paper" component={PaperScreen} />
-            <DiscoverStack.Screen name="Popover" component={PopoverScreen} />
-            <DiscoverStack.Screen name="Button" component={ButtonScreen} />
-            <DiscoverStack.Screen
-                name="SelectionControls"
-                component={SelectionControlsScreen}
-                options={{ headerTitle: "Selection Controls" }}
+            <ComponentsStack.Screen name="Discover" component={DiscoverScreen} />
+            <ComponentsStack.Screen
+                name="Component"
+                component={ComponentScreen}
+                options={({ route }) => ({
+                    title: ComponentName[route.params.name],
+                    ...(route.params.screenOptions ?? {}),
+                })}
             />
-            <DiscoverStack.Screen name="TextField" component={TextFieldScreen} />
-            <DiscoverStack.Screen name="Input" component={InputScreen} />
-            <DiscoverStack.Screen name="Autocomplete" component={AutocompleteScreen} />
-            <DiscoverStack.Screen name="Search" component={SearchScreen} />
-            <DiscoverStack.Screen name="Cell" component={CellScreen} />
-            <DiscoverStack.Screen name="NavigationCell" component={NavigationCellScreen} />
-            <DiscoverStack.Screen name="ButtonCell" component={ButtonCellScreen} />
-            <DiscoverStack.Screen name="SwitchCell" component={SwitchCellScreen} />
-            <DiscoverStack.Screen name="Accordion" component={AccordionScreen} />
-            <DiscoverStack.Screen name="Menu" component={MenuScreen} />
-            <DiscoverStack.Screen
-                name="ProgressIndicator"
-                options={{ title: "Progress Indicators" }}
-                component={ProgressIndicatorScreen}
-            />
-            <DiscoverStack.Screen name="Portal" component={PortalScreen} />
-            <DiscoverStack.Screen name="Dialog" component={DialogScreen} />
-            <DiscoverStack.Screen name="Environment" component={EnvironmentScreen} />
-            <DiscoverStack.Screen name="ErrorBoundary" component={ErrorBoundaryScreen} />
-        </DiscoverStack.Navigator>
+        </ComponentsStack.Navigator>
     );
 }
 
@@ -142,17 +110,17 @@ const BottomTab = createBottomTabNavigator<RootTabParamList>();
 function BottomTabNavigator() {
     return (
         <BottomTab.Navigator
-            initialRouteName="DiscoverTab"
+            initialRouteName="Components"
             screenOptions={{
                 tabBarLabelStyle: { fontFamily: "Equinor-Bold" },
                 headerRight: () => <GoToSettingsButton marginRight={8} />,
             }}
         >
             <BottomTab.Screen
-                name="DiscoverTab"
+                name="Components"
                 component={DiscoverNavigator}
                 options={{
-                    title: "Discover",
+                    title: "Components",
                     headerShown: false,
                     tabBarIcon: ({ color }) => (
                         <TabBarIcon name="binoculars" color={color as Color} />