+++ apps/chronicles/settings.ts
@@ -1,9 +1,8 @@
-import appJson from "./app.json"
-import { Platform } from "react-native"
+import appJson from "./app.json";
+import { Platform } from "react-native";
 
 export const getBuildNumber = (): string => {
-    const buildNumber = Platform.OS == "ios" ?
-        appJson.expo.ios.buildNumber as string :
-        appJson.expo.web.buildNumber as string;
+    const buildNumber =
+        Platform.OS == "ios" ? appJson.expo.ios.buildNumber : appJson.expo.web.buildNumber;
     return buildNumber;
-}
+};