+++ apps/chronicles/types/navigation.ts
@@ -1,0 +1,51 @@
+/**
+ * Learn more about using TypeScript with React Navigation:
+ * https://reactnavigation.org/docs/typescript/
+ */
+
+import { BottomTabScreenProps } from "@react-navigation/bottom-tabs";
+import { CompositeScreenProps, NavigatorScreenParams } from "@react-navigation/native";
+import {
+    NativeStackNavigationOptions,
+    NativeStackScreenProps,
+} from "@react-navigation/native-stack";
+import { ComponentType } from "./components";
+
+export type RootStackParamList = {
+    Root: NavigatorScreenParams<RootTabParamList> | undefined;
+    Login: undefined;
+    Modal: undefined;
+    NotFound: undefined;
+};
+
+export type ComponentsStackParamList = {
+    Discover: undefined;
+    Component: {
+        name: ComponentType;
+        screenOptions?: Partial<NativeStackNavigationOptions>;
+    };
+};
+
+export type DiscoverStackScreenProps<Screen extends keyof ComponentsStackParamList> =
+    NativeStackScreenProps<ComponentsStackParamList, Screen>;
+
+export type RootStackScreenProps<Screen extends keyof RootStackParamList> = NativeStackScreenProps<
+    RootStackParamList,
+    Screen
+>;
+
+export type RootTabParamList = {
+    Components: undefined;
+    IconsTab: undefined;
+    DrawTab: undefined;
+    SignTab: undefined;
+    Settings: undefined;
+};
+
+export type RootTabScreenProps<Screen extends keyof RootTabParamList> = CompositeScreenProps<
+    BottomTabScreenProps<RootTabParamList, Screen>,
+    CompositeScreenProps<
+        NativeStackScreenProps<RootStackParamList>,
+        NativeStackScreenProps<ComponentsStackParamList>
+    >
+>;