+++ packages/components/src/components/Cell/index.ts
@@ -3,7 +3,8 @@ import { NavigationCell, NavigationCellProps } from "./NavigationCell";
 import { Cell as _Cell, CellProps } from "./Cell";
 import { CellSwipeItemProps } from "./types";
 import { ButtonCell, ButtonCellProps } from "./ButtonCell";
-import { SwitchCell,SwitchCellProps } from "./SwitchCell";
+import { SwitchCell, SwitchCellProps } from "./SwitchCell";
+import { CellSwipeItem } from "./CellSwipeItem";
 
 type CellFamily = typeof _Cell & {
     /**
@@ -28,4 +29,12 @@ Cell.Navigation = NavigationCell;
 Cell.Button = ButtonCell;
 Cell.Switch = SwitchCell;
 
-export { Cell, CellProps, CellGroupProps, NavigationCellProps, CellSwipeItemProps, SwitchCellProps, ButtonCellProps };
+export { Cell, CellSwipeItem };
+export type {
+    CellProps,
+    CellGroupProps,
+    NavigationCellProps,
+    CellSwipeItemProps,
+    SwitchCellProps,
+    ButtonCellProps,
+};