+++ packages/core/src/components/createCoreStackNavigator.tsx
@@ -1,12 +1,13 @@
-import { CoreStackParamListBase, MadConfig } from "../types";
+import {CoreStackParamListBase, MadConfig} from "../types";
 import { createNativeStackNavigator } from "./navigation";
 import { createMadCoreNavigator } from "../utils/createMadCoreNavigator";
-import { setConfig } from "../store/mad-config";
+import {setConfig} from "../store/mad-config";
+import { ParamListBase } from "@react-navigation/native";
 
-export const createCoreStackNavigator = (config: MadConfig) => {
+export const createCoreStackNavigator = <T extends ParamListBase>(config: MadConfig) => {
     setConfig(config);
-    const Stack = createNativeStackNavigator<CoreStackParamListBase>();
-    const Navigator = createMadCoreNavigator(Stack, config);
+    const Stack = createNativeStackNavigator<CoreStackParamListBase & T>();
+    const Navigator = createMadCoreNavigator(Stack);
 
     return { ...Stack, Navigator };
 };