+++ packages/core/src/components/screens/CreateIncidentScreen.tsx
@@ -1,6 +1,0 @@
-import React from "react";
-import { Typography } from "@equinor/mad-components";
-
-export const CreateIncidentScreen = () => {
-    return <Typography>Not implemented yet</Typography>;
-};